refactor: extract buildUpdateInput shared by Query + TransactionQuery (#4)
Both classes built the same UpdateItem input shape from scratch
(attr names/values, SET assignments, the attribute_exists(#pk)
condition). Moved that to filterHelpers.buildUpdateInput so the
two callers reduce to 'send the input now' (Query) vs 'buffer the
input for transactional commit' (TransactionQuery), which is the
real behavioral difference.

Insert and remove are intentionally kept separate: Query.insert is
single-row with attribute_not_exists, Query.remove uses BatchWrite
when given an id list, while the transactional variants stage to
TransactWriteItems and forbid batch removes — different semantics,
not just different sinks.
